# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RadarTcell API
  version: "1.0.0"
  description: OpenAPI spec for RadarTcell backend (public + admin CRUD)
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/admin/login:
    post:
      tags: [Admin.Auth]
      summary: Admin login
      description: Возвращает JWT токен администратора
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/me:
    get:
      tags: [Admin.Auth]
      summary: Current admin user
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------- ADMIN CRUD: TRENDS ----------------
  /api/admin/trends:
    post:
      tags: [Admin.Trends]
      summary: Create trend
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTrendCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Trends]
      summary: List trends
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminTrend" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/trends/{slug}:
    put:
      tags: [Admin.Trends]
      summary: Update trend
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTrendUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [Admin.Trends]
      summary: Delete trend
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: No content }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Trends]
      summary: Get trend
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminTrend" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ---------------- ADMIN CRUD: TAGS ----------------
  /api/admin/tags:
    post:
      tags: [Admin.Tags]
      summary: Create tag
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTagCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Tags]
      summary: List tags
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TagListItem" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/admin/tags/{slug}:
    put:
      tags: [Admin.Tags]
      summary: Update tag
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTagUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [Admin.Tags]
      summary: Delete tag
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: No content }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Tags]
      summary: Get tag
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagListItem" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ---------------- ADMIN CRUD: ORGANIZATIONS ----------------
  /api/admin/organizations:
    post:
      tags: [Admin.Organizations]
      summary: Create organization
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminOrganizationCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Organizations]
      summary: List organizations
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrganizationItem" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/admin/organizations/{slug}:
    put:
      tags: [Admin.Organizations]
      summary: Update organization
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminOrganizationUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [Admin.Organizations]
      summary: Delete organization
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: No content }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Organizations]
      summary: Get organization
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationItem" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ---------------- ADMIN CRUD: METRICS ----------------
...
  /api/admin/technologies:
    post:
      tags: [Admin.Technologies]
      summary: Create technology
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTechnologyCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Technologies]
      summary: List technologies
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminTechnology" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/admin/technologies/{slug}:
    put:
      tags: [Admin.Technologies]
      summary: Update technology
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTechnologyUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminIdSlugResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [Admin.Technologies]
      summary: Delete technology
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: No content }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.Technologies]
      summary: Get technology
      security: [{ BearerAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminTechnology" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ---------------- ADMIN CRUD: SDGS ----------------
  /api/admin/sdgs:
    post:
      tags: [Admin.SDGs]
      summary: Create SDG
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSDGCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminCodeIdResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.SDGs]
      summary: List SDGs
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminSDG" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/admin/sdgs/{code}:
    put:
      tags: [Admin.SDGs]
      summary: Update SDG
      security: [{ BearerAuth: [] }]
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, example: "SDG 09" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSDGUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminCodeResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [Admin.SDGs]
      summary: Delete SDG
      security: [{ BearerAuth: [] }]
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, example: "SDG 09" }
      responses:
        "204": { description: No content }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Admin.SDGs]
      summary: Get SDG
      security: [{ BearerAuth: [] }]
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, example: "SDG 09" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSDG" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ---------------- PUBLIC: CATALOG ----------------
  /api/trends:
    get:
      tags: [Catalog]
      summary: List trends
      parameters:
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrendListResponse" }
  /api/sdgs:
    get:
      tags: [Catalog]
      summary: List SDGs
      parameters:
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SDGListResponse" }
  /api/tags:
    get:
      tags: [Catalog]
      summary: List tags
      parameters:
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagListResponse" }
  /api/organizations:
    get:
      tags: [Catalog]
      summary: List organizations
      parameters:
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationListResponse" }
...
  /api/technologies:
    get:
      tags: [Technology]
      summary: List technologies
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: trl_min
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 9 }
        - name: trl_max
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 9 }
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
          description: "Sort by name or any orderable metric field_key"
        - name: order
          in: query
          required: false
          schema: { type: string, enum: [asc, desc] }
        - name: highlight
          in: query
          required: false
          description: Repeatable, e.g. highlight=tag:ml&highlight=trend:ai
          style: form
          explode: true
          schema:
            type: array
            items: { type: string }
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TechnologyListResponse" }
  /api/technologies/{slug}:
    get:
      tags: [Technology]
      summary: Get technology by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TechnologyDetailsResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/trends/{slug}/technologies:
    get:
      tags: [Technology]
      summary: List technologies by trend
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TechnologyListResponse" }
  /api/sdgs/{code}/technologies:
    get:
      tags: [Technology]
      summary: List technologies by SDG code
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, example: "SDG 09" }
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TechnologyListResponse" }
  /api/tags/{slug}/technologies:
    get:
      tags: [Technology]
      summary: List technologies by tag
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TechnologyListResponse" }
  /api/organizations/{slug}/technologies:
    get:
      tags: [Technology]
      summary: List technologies by organization
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: locale
          in: query
          required: false
          schema: { type: string, example: "ru" }
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TechnologyListResponse" }

  # ---------------- PUBLIC: PREFERENCES ----------------
  /api/preferences:
    post:
      tags: [Preferences]
      summary: Save user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreferencesSaveRequest" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreferencesSaveResponse" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/preferences/{user_id}:
    get:
      tags: [Preferences]
      summary: Get user preferences
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreferencesGetResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string, example: not found }

    # -------- ADMIN AUTH --------
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, example: admin123 }
    AdminLoginResponse:
      type: object
      required: [token]
      properties:
        token: { type: string, example: "eyJhbGciOi..." }
    AdminMeResponse:
      type: object
      required: [role, user]
      properties:
        role: { type: string, example: admin }
        user: { type: string, example: admin }
...
    AdminIdSlugResponse:
      type: object
      required: [id, slug]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
...
    AdminTrendCreate:
      type: object
      required: [slug, name, order_index]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        image_url: { type: string, nullable: true }
        order_index: { type: integer }
...
    AdminTrend:
      type: object
      required: [id, slug, name, order_index]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        image_url: { type: string, nullable: true }
        order_index: { type: integer }
...
    AdminTagUpdate:
      type: object
      required: [title, category]
      properties:
        title: { type: string }
        category: { type: string }
        description: { type: string, nullable: true }
...